requirements.txt
pydantic-ai
openai
streamlit
python-dotenv
logfire